feat(237-05): extract OGC record asset helpers
- move asset builders, theme/time helpers, and dataset_to_ogc_record into service_records.py

- preserve build_assets, _build_stac_assets, and dataset_to_ogc_record through the public search facade
